Title: Add support for analyzing all Python files in a directory
Details:
Extended the functionality of the python_code_analyzer.py script to allow analysis of all Python files in a given directory. This includes parsing each file and extracting function/method details, dependencies, and entry points. To enable this, argparse was utilized to add a new '--directory' option. Users can now specify either a single file with the '--file' option or a directory with the '--directory' option.